Very Close
It can see if a guess is correct or wrong however, we need to get it to display the guessed letter along with the not guessed "_".  If anyone can do this feel free to play around with it. Just make sure you have a backup in case it doesn't work.